refactor: client file structure for RTK
