+++ client/app/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Navbar from './components/Navbar';
+import Navbar from '../features/navbar/Navbar';
 import AppRoutes from './AppRoutes';
 
 const App = () => {
