+++ client/app/AppRoutes.js
@@ -1,8 +1,8 @@
 import React, { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { Route, Routes } from 'react-router-dom';
-import AuthForm from './components/AuthForm';
-import Home from './components/Home';
+import AuthForm from '../features/auth/AuthForm';
+import Home from '../features/home/Home';
 import { me } from './store';
 
 /**
