+++ client/features/auth/AuthForm.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { authenticate } from '../store';
+import { authenticate } from '../../app/store';
 
 /**
   The AuthForm component can be used for Login or Sign Up.
