+++ client/index.js
@@ -1,11 +1,9 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
-import store from './store';
-import App from './App';
-import { MemoryRouter, BrowserRouter } from 'react-router-dom';
-
-const Router = process.env.NODE_ENV === 'test' ? MemoryRouter : BrowserRouter;
+import store from './app/store.js';
+import App from './app/App';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 const root = createRoot(document.getElementById('app'));
